# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Loop API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface ErrorResponse {
  message: string;
}

export type Cadence = (typeof Cadence)[keyof typeof Cadence];

export const Cadence = {
  one_time: "one_time",
  weekly: "weekly",
  monthly: "monthly",
  quarterly: "quarterly",
} as const;

export interface SignupBody {
  /** @minLength 1 */
  fullName: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: string;
  fullName: string;
  email: string;
}

export interface AuthResponse {
  user: User;
}

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  notes?: string | null;
  cadence: Cadence;
  followUpDate?: string | null;
}

export interface Client {
  id: string;
  name: string;
  notes?: string | null;
  cadence: Cadence;
  followUpDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ClientWithNext = Client & {
  nextFollowUpDate: string | null;
};

export interface FollowUpWithClient {
  id: string;
  clientId: string;
  clientName: string;
  notesSnapshot: string | null;
  cadence: Cadence;
  scheduledDate: string;
  completed: boolean;
}

export interface CompleteFollowUpResponse {
  id: string;
  completed: boolean;
  nextFollowUpDate: string | null;
}