# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Loop API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: clients
  - name: followUps
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "200":
          description: Signed up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      responses:
        "200":
          description: Clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientWithNext"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientWithNext"
  /clients/{id}:
    put:
      operationId: updateClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientWithNext"
    delete:
      operationId: deleteClient
      tags: [clients]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /follow-ups/today:
    get:
      operationId: getTodayFollowUps
      tags: [followUps]
      summary: Today's pending follow-ups
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FollowUpWithClient"
  /follow-ups/{id}/complete:
    patch:
      operationId: completeFollowUp
      tags: [followUps]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteFollowUpResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    Cadence:
      type: string
      enum: [one_time, weekly, monthly, quarterly]
    SignupBody:
      type: object
      properties:
        fullName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [fullName, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    User:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
      required: [id, fullName, email]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required: [user]
    ClientInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        notes:
          type: string
          nullable: true
        cadence:
          $ref: "#/components/schemas/Cadence"
        followUpDate:
          type: string
          format: date
          nullable: true
      required: [name, cadence]
    Client:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        notes:
          type: string
          nullable: true
        cadence:
          $ref: "#/components/schemas/Cadence"
        followUpDate:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, cadence, createdAt, updatedAt]
    ClientWithNext:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            nextFollowUpDate:
              type: string
              format: date
              nullable: true
          required: [nextFollowUpDate]
    FollowUpWithClient:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        notesSnapshot:
          type: string
          nullable: true
        cadence:
          $ref: "#/components/schemas/Cadence"
        scheduledDate:
          type: string
          format: date
        completed:
          type: boolean
      required: [id, clientId, clientName, notesSnapshot, cadence, scheduledDate, completed]
    CompleteFollowUpResponse:
      type: object
      properties:
        id:
          type: string
        completed:
          type: boolean
        nextFollowUpDate:
          type: string
          format: date
          nullable: true
      required: [id, completed, nextFollowUpDate]