# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Loop API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create an account
 */

export const signupBodyPasswordMin = 8;

export const SignupBody = zod.object({
  fullName: zod.string().min(1),
  email: zod.string().email(),
  password: zod.string().min(signupBodyPasswordMin),
});

export const SignupResponse = zod.object({
  user: zod.object({
    id: zod.string(),
    fullName: zod.string(),
    email: zod.string(),
  }),
});

/**
 * @summary Log in
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.string(),
    fullName: zod.string(),
    email: zod.string(),
  }),
});

/**
 * @summary Log out
 */
export const LogoutResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.string(),
  fullName: zod.string(),
  email: zod.string(),
});

/**
 * @summary List clients
 */
export const ListClientsResponseItem = zod
  .object({
    id: zod.string(),
    name: zod.string(),
    notes: zod.string().nullish(),
    cadence: zod.enum(["one_time", "weekly", "monthly", "quarterly"]),
    followUpDate: zod.coerce.date().nullish(),
    createdAt: zod.coerce.date(),
    updatedAt: zod.coerce.date(),
  })
  .and(
    zod.object({
      nextFollowUpDate: zod.coerce.date().nullable(),
    }),
  );
export const ListClientsResponse = zod.array(ListClientsResponseItem);

/**
 * @summary Create a client
 */

export const CreateClientBody = zod.object({
  name: zod.string().min(1),
  notes: zod.string().nullish(),
  cadence: zod.enum(["one_time", "weekly", "monthly", "quarterly"]),
  followUpDate: zod.coerce.date().nullish(),
});

export const CreateClientResponse = zod
  .object({
    id: zod.string(),
    name: zod.string(),
    notes: zod.string().nullish(),
    cadence: zod.enum(["one_time", "weekly", "monthly", "quarterly"]),
    followUpDate: zod.coerce.date().nullish(),
    createdAt: zod.coerce.date(),
    updatedAt: zod.coerce.date(),
  })
  .and(
    zod.object({
      nextFollowUpDate: zod.coerce.date().nullable(),
    }),
  );

export const UpdateClientParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateClientBody = zod.object({
  name: zod.string().min(1),
  notes: zod.string().nullish(),
  cadence: zod.enum(["one_time", "weekly", "monthly", "quarterly"]),
  followUpDate: zod.coerce.date().nullish(),
});

export const UpdateClientResponse = zod
  .object({
    id: zod.string(),
    name: zod.string(),
    notes: zod.string().nullish(),
    cadence: zod.enum(["one_time", "weekly", "monthly", "quarterly"]),
    followUpDate: zod.coerce.date().nullish(),
    createdAt: zod.coerce.date(),
    updatedAt: zod.coerce.date(),
  })
  .and(
    zod.object({
      nextFollowUpDate: zod.coerce.date().nullable(),
    }),
  );

export const DeleteClientParams = zod.object({
  id: zod.coerce.string(),
});

export const DeleteClientResponse = zod.object({
  ok: zod.boolean(),
});

/**
 * @summary Today's pending follow-ups
 */
export const GetTodayFollowUpsResponseItem = zod.object({
  id: zod.string(),
  clientId: zod.string(),
  clientName: zod.string(),
  notesSnapshot: zod.string().nullable(),
  cadence: zod.enum(["one_time", "weekly", "monthly", "quarterly"]),
  scheduledDate: zod.coerce.date(),
  completed: zod.boolean(),
});
export const GetTodayFollowUpsResponse = zod.array(
  GetTodayFollowUpsResponseItem,
);

export const CompleteFollowUpParams = zod.object({
  id: zod.coerce.string(),
});

export const CompleteFollowUpResponse = zod.object({
  id: zod.string(),
  completed: zod.boolean(),
  nextFollowUpDate: zod.coerce.date().nullable(),
});