# lib/api-zod/src/generated/types/cadence.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Loop API specification
 * OpenAPI spec version: 0.1.0
 */

export type Cadence = (typeof Cadence)[keyof typeof Cadence];

export const Cadence = {
  one_time: "one_time",
  weekly: "weekly",
  monthly: "monthly",
  quarterly: "quarterly",
} as const;