# lib/api-zod/src/generated/types/clientInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Loop API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Cadence } from "./cadence";

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  notes?: string | null;
  cadence: Cadence;
  followUpDate?: Date | null;
}