# lib/api-zod/src/generated/types/clientWithNext.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Loop API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Client } from "./client";

export type ClientWithNext = Client & {
  nextFollowUpDate: Date | null;
};