# lib/api-zod/src/generated/types/completeFollowUpResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Loop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CompleteFollowUpResponse {
  id: string;
  completed: boolean;
  nextFollowUpDate: Date | null;
}