# lib/api-zod/src/generated/types/followUpWithClient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Loop API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Cadence } from "./cadence";

export interface FollowUpWithClient {
  id: string;
  clientId: string;
  clientName: string;
  notesSnapshot: string | null;
  cadence: Cadence;
  scheduledDate: Date;
  completed: boolean;
}