# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Loop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SignupBody {
  /** @minLength 1 */
  fullName: string;
  email: string;
  /** @minLength 8 */
  password: string;
}