# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Loop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  fullName: string;
  email: string;
}